requirements.txt
fastapi
uvicorn
weaviate-client
langchain
openai
tiktoken
python-dotenv
httpx
google-genai
sentence-transformers